refactor: reorganize code structure for better modularity
- Move pkg/models to pkg/tdh/models to keep all tdh code together
- Split todo.go into focused modules:
  - operations.go: Core operations (Find, Swap, RemoveAtIndex, RemoveFinishedTodos)
  - filters.go: Filtering functions (with both mutating and non-mutating versions)
  - output.go: Display/formatting functions (MakeOutput with hashtag highlighting)
- Update all imports to use new models path
- Add non-mutating filter functions (FilterPending, FilterDone) alongside existing ones

This improves code organization by grouping related functionality and makes
the codebase more maintainable and easier to navigate.
