+++ pkg/tdh/store/memory_store.go
@@ -3,7 +3,7 @@ package store
 import (
 	"os"
 
-	"github.com/arthur-debert/tdh/pkg/models"
+	"github.com/arthur-debert/tdh/pkg/tdh/models"
 )
 
 // MemoryStore implements the Store interface for testing purposes.
