+++ pkg/tdh/store/store_test.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/arthur-debert/tdh/pkg/models"
+	"github.com/arthur-debert/tdh/pkg/tdh/models"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
